+++ frontend/src/data/diagnoses.ts
@@ -1,0 +1,90 @@
+/**
+ * Common Diagnoses for GI Surgery
+ * Categorized by type for easier selection in episode creation
+ */
+
+export const commonDiagnoses = {
+  malignant: [
+    'Colorectal Carcinoma',
+    'Rectal Carcinoma',
+    'Sigmoid Carcinoma',
+    'Caecal Carcinoma',
+    'Oesophageal Carcinoma',
+    'Gastric Carcinoma',
+    'Pancreatic Carcinoma',
+    'Hepatocellular Carcinoma',
+    'Cholangiocarcinoma',
+    'Small Bowel Tumour',
+    'Appendiceal Tumour',
+    'Carcinoid Tumour',
+    'GIST (Gastrointestinal Stromal Tumour)',
+    'Metastatic Disease to Liver',
+    'Peritoneal Carcinomatosis'
+  ],
+  inflammatory: [
+    'Crohn\'s Disease',
+    'Ulcerative Colitis',
+    'Diverticulitis',
+    'Diverticular Disease',
+    'Acute Appendicitis',
+    'Perforated Appendicitis',
+    'Perianal Sepsis',
+    'Perianal Abscess',
+    'Perianal Fistula',
+    'Anal Fissure',
+    'Pilonidal Sinus',
+    'Cholecystitis (Acute)',
+    'Cholecystitis (Chronic)',
+    'Pancreatitis (Chronic)',
+    'Ischaemic Colitis'
+  ],
+  benign: [
+    'Haemorrhoids',
+    'Rectal Prolapse',
+    'Gallstones (Cholelithiasis)',
+    'Common Bile Duct Stones (Choledocholithiasis)',
+    'Gastro-Oesophageal Reflux Disease (GORD)',
+    'Hiatus Hernia',
+    'Paraoesophageal Hernia',
+    'Peptic Ulcer Disease',
+    'Gastric Ulcer (Perforated)',
+    'Duodenal Ulcer (Perforated)',
+    'Bowel Obstruction',
+    'Small Bowel Obstruction',
+    'Large Bowel Obstruction',
+    'Adhesional Obstruction',
+    'Sigmoid Volvulus',
+    'Caecal Volvulus',
+    'Intestinal Perforation',
+    'Colonic Perforation',
+    'Faecal Impaction',
+    'Rectal Bleeding',
+    'Lower GI Bleeding'
+  ],
+  hernia: [
+    'Inguinal Hernia',
+    'Femoral Hernia',
+    'Umbilical Hernia',
+    'Paraumbilical Hernia',
+    'Epigastric Hernia',
+    'Incisional Hernia',
+    'Ventral Hernia',
+    'Spigelian Hernia',
+    'Strangulated Hernia',
+    'Obstructed Hernia'
+  ],
+  other: [
+    'Morbid Obesity',
+    'Achalasia',
+    'Barrett\'s Oesophagus',
+    'Oesophageal Stricture',
+    'Gastric Outlet Obstruction',
+    'Splenic Injury/Rupture',
+    'Mesenteric Ischaemia',
+    'Intestinal Ischaemia',
+    'Anastomotic Leak',
+    'Stoma Complications',
+    'Ventral Hernia',
+    'Enterocutaneous Fistula'
+  ]
+}
