+++ frontend/src/data/procedures.ts
@@ -1,0 +1,248 @@
+/**
+ * Surgical Procedure Data
+ * Maps procedure names to ICD-10 and OPCS-4 codes
+ * Used for episode creation and treatment documentation
+ */
+
+export const procedureToICD10: Record<string, string[]> = {
+  'Oesophagectomy (Ivor Lewis)': ['0DT50ZZ', '0DT54ZZ'],
+  'Oesophagectomy (Transhiatal)': ['0DT50ZZ', '0DT54ZZ'],
+  'Oesophagectomy (McKeown Three-Stage)': ['0DT50ZZ', '0DT54ZZ'],
+  'Oesophagogastrectomy': ['0DT50ZZ', '0DT60ZZ'],
+  'Total Gastrectomy': ['0DT60ZZ', '0DT64ZZ'],
+  'Partial Gastrectomy': ['0DB60ZZ', '0DB64ZZ'],
+  'Subtotal Gastrectomy': ['0DB60ZZ', '0DB64ZZ'],
+  'Sleeve Gastrectomy': ['0DB64Z3', '0DB60Z3'],
+  'Gastric Bypass (Roux-en-Y)': ['0D160ZA', '0D164ZA'],
+  'Gastrojejunostomy': ['0D160ZA', '0D164ZA'],
+  'Vagotomy (Truncal)': ['0D8X0ZZ', '0D8X4ZZ'],
+  'Vagotomy (Selective)': ['0D8X0ZZ', '0D8X4ZZ'],
+  'Fundoplication (Nissen)': ['0DU64JZ', '0DU60JZ'],
+  'Fundoplication (Toupet)': ['0DU64JZ', '0DU60JZ'],
+  'Fundoplication (Dor)': ['0DU64JZ', '0DU60JZ'],
+  'Pyloroplasty': ['0D970ZZ', '0D974ZZ'],
+  'Gastrostomy (PEG)': ['0DH60UZ', '0DH63UZ'],
+  'Hiatal Hernia Repair': ['0DU64JZ', '0DU60JZ'],
+  'Paraoesophageal Hernia Repair': ['0DU64JZ', '0DU60JZ'],
+  'Whipple Procedure (Pancreaticoduodenectomy)': ['0FTG0ZZ', '0FT90ZZ', '0DT90ZZ'],
+  'Distal Pancreatectomy': ['0FTG0ZZ', '0FTG4ZZ'],
+  'Splenectomy': ['07TP0ZZ', '07TP4ZZ'],
+  'Cholecystectomy': ['0FT40ZZ', '0FT44ZZ'],
+  'Common Bile Duct Exploration': ['0F9C0ZZ', '0F9C4ZZ'],
+  'Hepatectomy (Right)': ['0FT00ZZ', '0FT04ZZ'],
+  'Hepatectomy (Left)': ['0FT10ZZ', '0FT14ZZ'],
+  'Hepatectomy (Partial/Wedge)': ['0FTB0ZZ', '0FTB4ZZ'],
+  'Liver Resection (Segmentectomy)': ['0FTB0ZZ', '0FTB4ZZ'],
+  'Appendicectomy': ['0DTJ0ZZ', '0DTJ4ZZ'],
+  'Right Hemicolectomy': ['0DTF0ZZ', '0DTF4ZZ'],
+  'Extended Right Hemicolectomy': ['0DTF0ZZ', '0DTE0ZZ'],
+  'Left Hemicolectomy': ['0DTG0ZZ', '0DTG4ZZ'],
+  'Sigmoid Colectomy': ['0DTN0ZZ', '0DTN4ZZ'],
+  'Transverse Colectomy': ['0DTL0ZZ', '0DTL4ZZ'],
+  'Total Colectomy': ['0DTE0ZZ', '0DTE4ZZ'],
+  'Total Colectomy with Ileorectal Anastomosis': ['0DTE0ZZ', '0D1E0ZA'],
+  'Proctocolectomy with Ileostomy': ['0DTE0ZZ', '0DTP0ZZ'],
+  'Proctocolectomy with IPAA (J-Pouch)': ['0DTE0ZZ', '0D1E0ZC'],
+  'Anterior Resection': ['0DTP0ZZ', '0DTP4ZZ'],
+  'Low Anterior Resection': ['0DTP0ZZ', '0DTP4ZZ'],
+  'Ultra-Low Anterior Resection': ['0DTP0ZZ', '0DTP4ZZ'],
+  'Abdominoperineal Resection (APR)': ['0DTQ0ZZ', '0DTQ4ZZ'],
+  'Hartmann\'s Procedure': ['0DTN0ZZ', '0D1N0Z4'],
+  'Hartmann\'s Reversal': ['0D1N0ZA', '0D1N4ZA'],
+  'Ileostomy Formation (End)': ['0D1B0Z4', '0D1B4Z4'],
+  'Ileostomy Formation (Loop)': ['0D1B0Z4', '0D1B4Z4'],
+  'Colostomy Formation (End)': ['0D1L0Z4', '0D1N0Z4'],
+  'Colostomy Formation (Loop)': ['0D1L0Z4', '0D1N0Z4'],
+  'Stoma Reversal': ['0DQB0ZZ', '0DQN0ZZ'],
+  'Small Bowel Resection': ['0DTB0ZZ', '0DTB4ZZ'],
+  'Stricturoplasty': ['0D8B0ZZ', '0D8B4ZZ'],
+  'Adhesiolysis': ['0DN80ZZ', '0DN84ZZ'],
+  'Haemorrhoidectomy': ['0DBQ0ZZ', '0DBQ4ZZ'],
+  'Stapled Haemorrhoidopexy (PPH)': ['0DUQ0ZZ', '0DUQ4ZZ'],
+  'Lateral Internal Sphincterotomy': ['0D8R0ZZ', '0D8R4ZZ'],
+  'Fistulotomy': ['0D9Q0ZZ', '0D9Q4ZZ'],
+  'Fistula Repair (Advancement Flap)': ['0DUQ0ZZ', '0DUQ4ZZ'],
+  'Seton Placement': ['0DHQ0ZZ', '0DHQ4ZZ'],
+  'Pilonidal Sinus Excision': ['0HBU0ZZ', '0HBU4ZZ'],
+  'Perianal Abscess Drainage': ['0D9Q0ZZ', '0D9Q4ZZ'],
+  'Rectal Prolapse Repair (Perineal)': ['0DSQ0ZZ', '0DSQ4ZZ'],
+  'Rectopexy (Ventral Mesh)': ['0DSP0ZZ', '0DSP4ZZ'],
+  'Rectopexy (Laparoscopic Suture)': ['0DSP4ZZ', '0DSP0ZZ'],
+  'STARR Procedure': ['0DBP0ZZ', '0DBP4ZZ'],
+  'Inguinal Hernia Repair (Open)': ['0YU60JZ', '0YU50JZ'],
+  'Inguinal Hernia Repair (Laparoscopic TEP)': ['0YU64JZ', '0YU54JZ'],
+  'Inguinal Hernia Repair (Laparoscopic TAPP)': ['0YU64JZ', '0YU54JZ'],
+  'Femoral Hernia Repair': ['0YU60JZ', '0YU64JZ'],
+  'Incisional Hernia Repair': ['0WUF0JZ', '0WUF4JZ'],
+  'Umbilical Hernia Repair': ['0WU90JZ', '0WU94JZ'],
+  'Paraumbilical Hernia Repair': ['0WU90JZ', '0WU94JZ'],
+  'Epigastric Hernia Repair': ['0WUF0JZ', '0WUF4JZ'],
+  'Ventral Hernia Repair': ['0WUF0JZ', '0WUF4JZ'],
+  'Spigelian Hernia Repair': ['0WUF0JZ', '0WUF4JZ'],
+  'Component Separation (Anterior)': ['0KUM0JZ', '0KUM4JZ'],
+  'Component Separation (Posterior - TAR)': ['0KUM0JZ', '0KUM4JZ']
+}
+
+export const procedureToOPCS: Record<string, string[]> = {
+  'Oesophagectomy (Ivor Lewis)': ['G01.1', 'G01.8'],
+  'Oesophagectomy (Transhiatal)': ['G01.1', 'G01.8'],
+  'Oesophagectomy (McKeown Three-Stage)': ['G01.1', 'G01.8'],
+  'Oesophagogastrectomy': ['G01.2', 'G02.1'],
+  'Total Gastrectomy': ['G02.1', 'G02.8'],
+  'Partial Gastrectomy': ['G02.2', 'G02.3'],
+  'Subtotal Gastrectomy': ['G02.2', 'G02.3'],
+  'Sleeve Gastrectomy': ['G28.2', 'G27.2'],
+  'Gastric Bypass (Roux-en-Y)': ['G28.1', 'G27.1'],
+  'Gastrojejunostomy': ['G10.2', 'G10.8'],
+  'Vagotomy (Truncal)': ['G41.1', 'G41.8'],
+  'Vagotomy (Selective)': ['G41.2', 'G41.3'],
+  'Fundoplication (Nissen)': ['G17.1', 'G17.8'],
+  'Fundoplication (Toupet)': ['G17.1', 'G17.8'],
+  'Fundoplication (Dor)': ['G17.1', 'G17.8'],
+  'Pyloroplasty': ['G08.1', 'G08.8'],
+  'Gastrostomy (PEG)': ['G43.1', 'G43.8'],
+  'Hiatal Hernia Repair': ['G17.1', 'G17.8'],
+  'Paraoesophageal Hernia Repair': ['G17.1', 'G17.8'],
+  'Whipple Procedure (Pancreaticoduodenectomy)': ['J03.1', 'J03.2'],
+  'Distal Pancreatectomy': ['J03.3', 'J03.5'],
+  'Splenectomy': ['T86.1', 'T86.2'],
+  'Cholecystectomy': ['J18.1', 'J18.5'],
+  'Common Bile Duct Exploration': ['J15.1', 'J15.8'],
+  'Hepatectomy (Right)': ['J02.1', 'J02.5'],
+  'Hepatectomy (Left)': ['J02.2', 'J02.5'],
+  'Hepatectomy (Partial/Wedge)': ['J02.8', 'J02.9'],
+  'Liver Resection (Segmentectomy)': ['J02.3', 'J02.4'],
+  'Appendicectomy': ['H01.1', 'H01.8'],
+  'Right Hemicolectomy': ['H05.1', 'H05.2'],
+  'Extended Right Hemicolectomy': ['H04.1', 'H05.1'],
+  'Left Hemicolectomy': ['H06.1', 'H06.2'],
+  'Sigmoid Colectomy': ['H08.1', 'H08.2'],
+  'Transverse Colectomy': ['H07.1', 'H07.2'],
+  'Total Colectomy': ['H04.1', 'H04.2'],
+  'Total Colectomy with Ileorectal Anastomosis': ['H04.1', 'H10.2'],
+  'Proctocolectomy with Ileostomy': ['H04.3', 'H04.8'],
+  'Proctocolectomy with IPAA (J-Pouch)': ['H04.4', 'H04.5'],
+  'Anterior Resection': ['H33.1', 'H33.2'],
+  'Low Anterior Resection': ['H33.3', 'H33.4'],
+  'Ultra-Low Anterior Resection': ['H33.5', 'H33.6'],
+  'Abdominoperineal Resection (APR)': ['H33.8', 'H33.9'],
+  'Hartmann\'s Procedure': ['H08.3', 'H08.4'],
+  'Hartmann\'s Reversal': ['G62.1', 'G69.1'],
+  'Ileostomy Formation (End)': ['G69.1', 'G69.2'],
+  'Ileostomy Formation (Loop)': ['G69.3', 'G69.4'],
+  'Colostomy Formation (End)': ['G62.1', 'G62.2'],
+  'Colostomy Formation (Loop)': ['G62.3', 'G62.4'],
+  'Stoma Reversal': ['G64.1', 'G71.1'],
+  'Small Bowel Resection': ['G63.1', 'G63.8'],
+  'Stricturoplasty': ['G61.1', 'G61.8'],
+  'Adhesiolysis': ['T93.1', 'T93.8'],
+  'Haemorrhoidectomy': ['H51.1', 'H51.8'],
+  'Stapled Haemorrhoidopexy (PPH)': ['H51.2', 'H51.9'],
+  'Lateral Internal Sphincterotomy': ['H52.1', 'H52.8'],
+  'Fistulotomy': ['H53.1', 'H53.2'],
+  'Fistula Repair (Advancement Flap)': ['H56.1', 'H56.8'],
+  'Seton Placement': ['H59.1', 'H59.8'],
+  'Pilonidal Sinus Excision': ['S56.1', 'S56.8'],
+  'Perianal Abscess Drainage': ['H54.1', 'H54.8'],
+  'Rectal Prolapse Repair (Perineal)': ['H42.1', 'H42.2'],
+  'Rectopexy (Ventral Mesh)': ['H41.1', 'H41.8'],
+  'Rectopexy (Laparoscopic Suture)': ['H41.2', 'H41.9'],
+  'STARR Procedure': ['H33.7', 'H35.8'],
+  'Inguinal Hernia Repair (Open)': ['T20.1', 'T20.2'],
+  'Inguinal Hernia Repair (Laparoscopic TEP)': ['T20.3', 'T20.4'],
+  'Inguinal Hernia Repair (Laparoscopic TAPP)': ['T20.5', 'T20.8'],
+  'Femoral Hernia Repair': ['T21.1', 'T21.2'],
+  'Incisional Hernia Repair': ['T27.1', 'T27.2'],
+  'Umbilical Hernia Repair': ['T22.1', 'T22.2'],
+  'Paraumbilical Hernia Repair': ['T22.3', 'T22.4'],
+  'Epigastric Hernia Repair': ['T23.1', 'T23.2'],
+  'Ventral Hernia Repair': ['T27.3', 'T27.4'],
+  'Spigelian Hernia Repair': ['T27.8', 'T27.9'],
+  'Component Separation (Anterior)': ['T28.1', 'T28.2'],
+  'Component Separation (Posterior - TAR)': ['T28.3', 'T28.4']
+}
+
+export const standardProcedures = {
+  upperGI: [
+    'Oesophagectomy (Ivor Lewis)',
+    'Oesophagectomy (Transhiatal)',
+    'Oesophagectomy (McKeown Three-Stage)',
+    'Oesophagogastrectomy',
+    'Total Gastrectomy',
+    'Partial Gastrectomy',
+    'Subtotal Gastrectomy',
+    'Sleeve Gastrectomy',
+    'Gastric Bypass (Roux-en-Y)',
+    'Gastrojejunostomy',
+    'Vagotomy (Truncal)',
+    'Vagotomy (Selective)',
+    'Fundoplication (Nissen)',
+    'Fundoplication (Toupet)',
+    'Fundoplication (Dor)',
+    'Pyloroplasty',
+    'Gastrostomy (PEG)',
+    'Hiatal Hernia Repair',
+    'Paraoesophageal Hernia Repair',
+    'Whipple Procedure (Pancreaticoduodenectomy)',
+    'Distal Pancreatectomy',
+    'Splenectomy',
+    'Cholecystectomy',
+    'Common Bile Duct Exploration',
+    'Hepatectomy (Right)',
+    'Hepatectomy (Left)',
+    'Hepatectomy (Partial/Wedge)',
+    'Liver Resection (Segmentectomy)'
+  ],
+  lowerGI: [
+    'Appendicectomy',
+    'Right Hemicolectomy',
+    'Extended Right Hemicolectomy',
+    'Left Hemicolectomy',
+    'Sigmoid Colectomy',
+    'Transverse Colectomy',
+    'Total Colectomy',
+    'Total Colectomy with Ileorectal Anastomosis',
+    'Proctocolectomy with Ileostomy',
+    'Proctocolectomy with IPAA (J-Pouch)',
+    'Anterior Resection',
+    'Low Anterior Resection',
+    'Ultra-Low Anterior Resection',
+    'Abdominoperineal Resection (APR)',
+    'Hartmann\'s Procedure',
+    'Hartmann\'s Reversal',
+    'Ileostomy Formation (End)',
+    'Ileostomy Formation (Loop)',
+    'Colostomy Formation (End)',
+    'Colostomy Formation (Loop)',
+    'Stoma Reversal',
+    'Small Bowel Resection',
+    'Stricturoplasty',
+    'Adhesiolysis',
+    'Haemorrhoidectomy',
+    'Stapled Haemorrhoidopexy (PPH)',
+    'Lateral Internal Sphincterotomy',
+    'Fistulotomy',
+    'Fistula Repair (Advancement Flap)',
+    'Seton Placement',
+    'Pilonidal Sinus Excision',
+    'Perianal Abscess Drainage',
+    'Rectal Prolapse Repair (Perineal)',
+    'Rectopexy (Ventral Mesh)',
+    'Rectopexy (Laparoscopic Suture)',
+    'STARR Procedure'
+  ],
+  hernia: [
+    'Inguinal Hernia Repair (Open)',
+    'Inguinal Hernia Repair (Laparoscopic TEP)',
+    'Inguinal Hernia Repair (Laparoscopic TAPP)',
+    'Femoral Hernia Repair',
+    'Incisional Hernia Repair',
+    'Umbilical Hernia Repair',
+    'Paraumbilical Hernia Repair',
+    'Epigastric Hernia Repair',
+    'Ventral Hernia Repair',
+    'Spigelian Hernia Repair',
+    'Component Separation (Anterior)',
+    'Component Separation (Posterior - TAR)'
+  ]
+}
